moved globals.py into src
